time zero proc tool based on mas amplitude of mean trace

+++ code/radar/__init__.py
@@ -86,7 +86,7 @@ class radar(object):
     # convert amplitude array to dB log scale
     def dBscale(self, dat):
         # convert to power
-        pow = np.power(dat, 2)
+        pow = np.power(dat.astype(np.float), 2)
         # mask zero-power values
         pow[pow == 0] = np.nan
         # dB it
